Update build and change theme to black and white

+++ client.lua
@@ -39,7 +39,7 @@ lib.onCache('ped', function(ped)
 	Utils.WeaponWheel()
 end)
 
-plyState:set('invBusy', true, false)
+plyState:set('invBusy', true, true)
 plyState:set('invHotkeys', false, false)
 plyState:set('canUseWeapons', false, false)
 
@@ -159,6 +159,8 @@ function client.openInventory(inv, data)
         return lib.notify({ id = 'inventory_player_access', type = 'error', description = locale('inventory_player_access') })
     end
 
+	TriggerEvent('ox_inventory:UpdatePlayerDamage')
+	
     local left, right, accessError
 
     if inv == 'player' and data ~= cache.serverId then
@@ -415,6 +417,7 @@ local function canUseItem(isAmmo)
 	and not lib.progressActive()
 	and not IsPedRagdoll(ped)
 	and not IsPedFalling(ped)
+    and not IsPedShooting(playerPed)
 end
 
 ---@param data table
@@ -432,8 +435,7 @@ local function useItem(data, cb, noAnim)
     end
 
 	if currentWeapon and currentWeapon.timer ~= 0 then
-        if IsPedShooting(playerPed) then return end
-        if currentWeapon.timer - GetGameTimer() > 100 then return end
+        if not currentWeapon.timer or currentWeapon.timer - GetGameTimer() > 100 then return end
 
         DisablePlayerFiring(cache.playerId, true)
     end
@@ -518,6 +520,8 @@ local function useSlot(slot, noAnim)
 			if IsCinematicCamRendering() then SetCinematicModeActive(false) end
 
 			if currentWeapon then
+                if not currentWeapon.timer or currentWeapon.timer ~= 0 then return end
+
 				local weaponSlot = currentWeapon.slot
 				currentWeapon = Weapon.Disarm(currentWeapon)
 
@@ -528,6 +532,7 @@ local function useSlot(slot, noAnim)
             SetCurrentPedWeapon(playerPed, data.hash, false)
 
             if data.hash ~= GetSelectedPedWeapon(playerPed) then
+                lib.print.info(('failed to equip %s (cause unknown)'):format(item.name))
                 return lib.notify({ type = 'error', description = locale('cannot_use', data.label) })
             end
 
@@ -746,7 +751,7 @@ local function registerCommands()
 		-- No storage or no glovebox
 		if (checkVehicle == 0 or checkVehicle == 2) or (not Vehicles.glovebox[vehicleClass] and not Vehicles.glovebox.models[vehicleHash]) then return end
 
-		local isOpen = client.openInventory('glovebox', { id = 'glove'..GetVehicleNumberPlateText(vehicle), netid = NetworkGetNetworkIdFromEntity(vehicle) })
+		local isOpen = client.openInventory('glovebox', { netid = NetworkGetNetworkIdFromEntity(vehicle) })
 
 		if isOpen then
 			currentInventory.entity = vehicle
@@ -1026,6 +1031,10 @@ local function onEnterDrop(point)
 	if not point.instance or point.instance == currentInstance and not point.entity then
 		local model = point.model or client.dropmodel
 
+        -- Prevent breaking inventory on invalid point.model instead use default client.dropmodel
+        if not IsModelValid(model) and not IsModelInCdimage(model) then
+            model = client.dropmodel
+        end
 		lib.requestModel(model)
 
 		local entity = CreateObject(model, point.coords.x, point.coords.y, point.coords.z, false, true, true)
@@ -1123,19 +1132,21 @@ local function setStateBagHandler(stateId)
 	AddStateBagChangeHandler('instance', stateId, function(_, _, value)
 		currentInstance = value
 
-		-- Iterate over known drops and remove any points in a different instance (ignoring no instance)
-		for dropId, point in pairs(client.drops) do
-			if point.instance then
-				if point.instance ~= value then
-					if point.entity then
-						Utils.DeleteEntity(point.entity)
-						point.entity = nil
-					end
+		if client.drops then
+			-- Iterate over known drops and remove any points in a different instance (ignoring no instance)
+			for dropId, point in pairs(client.drops) do
+				if point.instance then
+					if point.instance ~= value then
+						if point.entity then
+							Utils.DeleteEntity(point.entity)
+							point.entity = nil
+						end
 
-					point:remove()
-				else
-					-- Recreate the drop using data from the old point
-					createDrop(dropId, point)
+						point:remove()
+					else
+						-- Recreate the drop using data from the old point
+						createDrop(dropId, point)
+					end
 				end
 			end
 		end
@@ -1177,6 +1188,7 @@ RegisterNetEvent('ox_inventory:setPlayerInventory', function(currentDrops, inven
     ---@class PlayerData
     ---@field inventory table<number, SlotWithItem?>
     ---@field weight number
+    ---@field groups table<string, number>
 	PlayerData = player
 	PlayerData.id = cache.playerId
 	PlayerData.source = cache.serverId
@@ -1270,7 +1282,7 @@ RegisterNetEvent('ox_inventory:setPlayerInventory', function(currentDrops, inven
 
 		if point.isClosest and point.currentDistance < 1.2 then
 			if not hasTextUi then
-				hasTextUi = nil
+				hasTextUi = point
 				lib.showTextUI(point.message, uiOptions)
 			end
 
@@ -1291,7 +1303,7 @@ RegisterNetEvent('ox_inventory:setPlayerInventory', function(currentDrops, inven
 		end
 	end
 
-	for id, data in pairs(lib.load('data.licenses')) do
+	for id, data in pairs(lib.load('data.licenses') or {}) do
 		lib.points.new({
 			coords = data.coords,
 			distance = 16,
@@ -1322,7 +1334,7 @@ RegisterNetEvent('ox_inventory:setPlayerInventory', function(currentDrops, inven
 	})
 
 	PlayerData.loaded = true
-
+	TriggerEvent('ox_inventory:UpdatePlayerDamage')
 	lib.notify({ description = locale('inventory_setup') })
 	Shops.refreshShops()
 	Inventory.Stashes()
@@ -1347,19 +1359,21 @@ RegisterNetEvent('ox_inventory:setPlayerInventory', function(currentDrops, inven
 				playerCoords = GetEntityCoords(playerPed)
 
 				if currentInventory and not currentInventory.ignoreSecurityChecks then
+                    local maxDistance = (currentInventory.distance or currentInventory.type == 'stash' and 4.8 or 1.8) + 0.2
+
 					if currentInventory.type == 'otherplayer' then
 						local id = GetPlayerFromServerId(currentInventory.id)
 						local ped = GetPlayerPed(id)
 						local pedCoords = GetEntityCoords(ped)
 
-						if not id or #(playerCoords - pedCoords) > 1.8 or not (client.hasGroup(shared.police) or canOpenTarget(ped)) then
+						if not id or #(playerCoords - pedCoords) > maxDistance or not (client.hasGroup(shared.police) or canOpenTarget(ped)) then
 							client.closeInventory()
 							lib.notify({ id = 'inventory_lost_access', type = 'error', description = locale('inventory_lost_access') })
 						else
 							TaskTurnPedToFaceCoord(playerPed, pedCoords.x, pedCoords.y, pedCoords.z, 50)
 						end
 
-					elseif currentInventory.coords and (#(playerCoords - currentInventory.coords) > (currentInventory.distance or 2.0) or canOpenTarget(playerPed)) then
+					elseif currentInventory.coords and (#(playerCoords - currentInventory.coords) > maxDistance or canOpenTarget(playerPed)) then
 						client.closeInventory()
 						lib.notify({ id = 'inventory_lost_access', type = 'error', description = locale('inventory_lost_access') })
 					end
@@ -1389,6 +1403,7 @@ RegisterNetEvent('ox_inventory:setPlayerInventory', function(currentDrops, inven
 				end
 
 				if weaponHash ~= currentWeapon.hash then
+                    lib.print.info(('%s was forcibly unequipped (caused by game behaviour or another resource)'):format(currentWeapon.name))
 					currentWeapon = Weapon.Disarm(currentWeapon, true)
 				end
 			end
@@ -1520,8 +1535,8 @@ RegisterNetEvent('ox_inventory:setPlayerInventory', function(currentDrops, inven
 							while IsPedPlantingBomb(playerPed) do Wait(0) end
 
 							TriggerServerEvent('ox_inventory:updateWeapon', 'throw', nil, weapon.slot)
+							plyState:set('invBusy', false, true)
 
-							plyState.invBusy = false
 							currentWeapon = nil
 
 							RemoveWeaponFromPed(playerPed, weapon.hash)
@@ -1536,7 +1551,7 @@ RegisterNetEvent('ox_inventory:setPlayerInventory', function(currentDrops, inven
 		end
 	end)
 
-	plyState:set('invBusy', false, false)
+	plyState:set('invBusy', false, true)
 	plyState:set('invOpen', false, false)
 	plyState:set('invHotkeys', true, false)
 	plyState:set('canUseWeapons', true, false)
@@ -1645,7 +1660,12 @@ local function giveItemToTarget(serverId, slotId, count)
     end
 
     Utils.PlayAnim(0, 'mp_common', 'givetake1_a', 1.0, 1.0, 2000, 50, 0.0, 0, 0, 0)
-    TriggerServerEvent('ox_inventory:giveItem', slotId, serverId, count or 0)
+
+    local notification = lib.callback.await('ox_inventory:giveItem', false, slotId, serverId, count or 0)
+
+    if notification then
+        lib.notify({ type = 'error', description = locale(table.unpack(notification)) })
+    end
 end
 
 exports('giveItemToTarget', giveItemToTarget)
@@ -1884,4 +1904,92 @@ lib.callback.register('ox_inventory:getVehicleData', function(netid)
 	if entity then
 		return GetEntityModel(entity), GetVehicleClass(entity)
 	end
+end)
+
+bodypercent = { 
+	['HEAD'] = { percent = 0, bullets = 0, severity = false, broken = false, bleeding = false },
+	['NECK'] = { percent = 0, bullets = 0, severity = false, broken = false, bleeding = false },
+	['UPPER_BODY'] = { percent = 0, bullets = 0, severity = false, broken = false, bleeding = false },
+	['LOWER_BODY'] = { percent = 0, bullets = 0, severity = false, broken = false, bleeding = false },
+	['LARM'] = { percent = 0, bullets = 0, severity = false, broken = false, bleeding = false },
+	['RARM'] = { percent = 0, bullets = 0, severity = false, broken = false, bleeding = false },
+	['LHAND'] = { percent = 0, bullets = 0, severity = false, broken = false, bleeding = false },
+	['RHAND'] = { percent = 0, bullets = 0, severity = false, broken = false, bleeding = false },
+	['LLEG'] = { percent = 0, bullets = 0, severity = false, broken = false, bleeding = false },
+	['RLEG'] = { percent = 0, bullets = 0, severity = false, broken = false, bleeding = false },
+	['LFOOT'] = { percent = 0, bullets = 0, severity = false, broken = false, bleeding = false },
+	['RFOOT'] = { percent = 0, bullets = 0, severity = false, broken = false, bleeding = false },
+	['SPINE'] = { percent = 0, bullets = 0, severity = false, broken = false, bleeding = false },
+}
+
+RegisterNetEvent('ox_inventory:UpdatePlayerDamage', function(BodyParts)
+	local GetPlayerDamage = BodyParts or exports.BC_Wounding:GetPlayerDamage()
+	for k, v in pairs(GetPlayerDamage) do
+		if v.severity and not bodypercent[k].severity then
+			bodypercent[k].percent = bodypercent[k].percent + 40
+			bodypercent[k].severity = true
+		end
+
+		if not v.severity and bodypercent[k].severity then
+			bodypercent[k].severity = false
+			bodypercent[k].percent = bodypercent[k].percent - 40
+		end
+
+		if v.broken and not bodypercent[k].broken then
+			bodypercent[k].percent = bodypercent[k].percent + 20
+			bodypercent[k].broken = true
+		end
+
+		if not v.broken and bodypercent[k].broken then
+			bodypercent[k].broken = false
+			bodypercent[k].percent = bodypercent[k].percent - 20
+		end
+
+		if v.bleeding and not bodypercent[k].bleeding then
+			bodypercent[k].percent = bodypercent[k].percent + 10
+			bodypercent[k].bleeding = true
+		end
+
+		if not v.bleeding and bodypercent[k].bleeding then
+			bodypercent[k].bleeding = false
+			bodypercent[k].percent = bodypercent[k].percent - 10
+		end
+
+		if v.bullet and v.bullet ~= bodypercent[k].bullets then
+			if v.bullet > bodypercent[k].bullets then
+				local bulletCalc = math.floor(v.bullet * 10)
+				bodypercent[k].percent = bodypercent[k].percent + bulletCalc
+				bodypercent[k].bullets = v.bullet
+			else
+				local bulletCalc = math.floor(v.bullet * 10)
+				bodypercent[k].percent = bodypercent[k].percent - bulletCalc
+				bodypercent[k].bullets = v.bullet
+			end
+		end
+
+		-- Ensure the value does not exceed 100
+        if bodypercent[k].percent > 100 then
+            bodypercent[k].percent = 100
+		elseif bodypercent[k].percent < 0 then
+			bodypercent[k].percent = 0
+		end
+	end
+	SendNUIMessage({
+		action = 'DamageCall',
+		data = bodypercent
+	})
+end)
+
+RegisterNetEvent('ox_inventory:resetdamageforplayer', function()
+    for k, v in pairs(bodypercent) do
+        v.percent = 0
+        v.bullets = 0
+        v.severity = false
+        v.broken = false
+        v.bleeding = false
+    end
+    SendNUIMessage({
+        action = 'DamageCall',
+        data = bodypercent
+    })
 end)
